magento/adobe-stock-integration#1578: Move MediaGallery related modules to Magento core - code review fixes
