--- AdobeStockImageAdminUi/etc/config.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
-    <default>
-        <system>
-            <media_gallery>
-                <enabled>0</enabled>
-            </media_gallery>
-        </system>
-    </default>
-</config>
