+++ MediaContentSynchronization/etc/di.xml
@@ -14,7 +14,7 @@
             </argument>
         </arguments>
     </type>
-    <type name="Magento\MediaGallerySynchronization\Model\Consume">
+    <type name="Magento\MediaGallerySynchronization\Model\Consumer">
         <plugin name="synchronize_media_content"
                 type="Magento\MediaContentSynchronization\Plugin\SynchronizeMediaContent"/>
     </type>
