+++ MediaGallerySynchronization/etc/queue_consumer.xml
@@ -7,5 +7,5 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework-message-queue:etc/consumer.xsd">
     <consumer name="media.gallery.synchronization" queue="media.gallery.synchronization"
-              connection="db" handler="Magento\MediaGallerySynchronization\Model\Consume::execute"/>
+              connection="db" handler="Magento\MediaGallerySynchronization\Model\Consumer::execute"/>
 </config>
