+++ MediaGalleryUi/Model/AssetDetailsProvider/UsedIn.php
@@ -71,15 +71,11 @@ class UsedIn implements AssetDetailsProviderInterface
         foreach ($contentIdentities as $contentIdentity) {
             $entityId = $contentIdentity->getEntityId();
             $type = $this->contentTypes[$contentIdentity->getEntityType()] ?? $contentIdentity->getEntityType();
-            
+
             if (!isset($entityIds[$type])) {
                 $usedIn[$type] = 1;
             } elseif ($entityIds[$type]['entity_id'] !== $entityId) {
-                if (isset($usedIn[$type])) {
-                    $usedIn[$type] +=1;
-                } else {
-                    $usedIn[$type] = $usedIn[$type]+1;
-                }
+                    ++$usedIn[$type];
             }
             $entityIds[$type]['entity_id'] = $entityId;
         }
