+++ MediaGalleryUi/Ui/Component/ImageUploaderStandAlone.php
@@ -24,9 +24,10 @@ class ImageUploaderStandAlone extends ImageUploader
             array_replace_recursive(
                 (array) $this->getData('config'),
                 [
-                    // phpcs:disable Magento2.Files.LineLength, Generic.Files.LineLength
-                    'actionsPath' => 'standalone_media_gallery_listing.standalone_media_gallery_listing.media_gallery_columns.thumbnail_url',
-                    'directoriesPath' => 'standalone_media_gallery_listing.standalone_media_gallery_listing.media_gallery_directories',
+                    'actionsPath' => 'standalone_media_gallery_listing.standalone_media_gallery_listing' .
+                        '.media_gallery_columns.thumbnail_url',
+                    'directoriesPath' => 'standalone_media_gallery_listing.standalone_media_gallery_listing' .
+                        '.media_gallery_directories',
                     'messagesPath' => 'standalone_media_gallery_listing.standalone_media_gallery_listing.messages'
                 ]
             )
