+++ MediaGalleryUi/etc/config.xml
@@ -9,7 +9,7 @@
     <default>
         <system>
             <media_gallery>
-                <enabled>1</enabled>
+                <enabled>0</enabled>
             </media_gallery>
         </system>
     </default>
